Limit requesting EV identification (#1530)
If the EEBUS EVSE doesn't provide an identification we retry endlessly. The problem is that getting the identification from the EV is not instant and could take 1-2 minutes after it is connected. Also if there are network issues, the EVSE doesn't provide the identification it previously had.

So instead of requesting the identification value endlessly, we limit it to a maximum count.